rearrange code to match messages structure, part 1
