+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/Background.cs
@@ -1,0 +1,8 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Gherkin.CucumberMessages.Types
+{
+    //TODO
+}
