+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/Comment.cs
@@ -1,12 +1,14 @@
 ﻿using System.Runtime.Serialization;
+using Gherkin.Events.Args.Ast;
 
-namespace Gherkin.Events.Args.Ast
+namespace Gherkin.CucumberMessages.Types
 {
     public class Comment
     {
-        [DataMember(Name="text")]
-        public string Text { get; set; }
         [DataMember(Name = "location")]
         public Location Location { get; set; }
+
+        [DataMember(Name="text")]
+        public string Text { get; set; }
     }
 }
