+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/DocString.cs
@@ -1,20 +1,20 @@
-using System.Collections.Generic;
 using System.Runtime.Serialization;
+using Gherkin.Events.Args.Ast;
 
-namespace Gherkin.Events.Args.Ast
+namespace Gherkin.CucumberMessages.Types
 {
     public class DocString
     {
+        [DataMember(Name = "location")]
+        public Location Location { get; set; }
+
+        [DataMember(Name = "mediaType")]
+        public string MediaType { get; set; }
+
         [DataMember(Name = "content")]
         public string Content { get; set; }
 
         [DataMember(Name = "delimiter")]
         public string Delimiter { get; set; }
-
-        [DataMember(Name = "mediaType")]
-        public string MediaType { get; set; }
-
-        [DataMember(Name = "location")]
-        public Location Location { get; set; }
     }
 }
