+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/FeatureChild.cs
@@ -1,17 +1,17 @@
 ﻿using System.Runtime.Serialization;
+using Gherkin.Events.Args.Ast;
 
-namespace Gherkin.Events.Args.Ast
+namespace Gherkin.CucumberMessages.Types
 {
-    public class Children
+    public class FeatureChild
     {
-        [DataMember(Name = "background")]
-        public StepsContainer Background { get; set; }
         [DataMember(Name = "scenario")]
         public StepsContainer Scenario { get; set; }
 
+        [DataMember(Name = "background")]
+        public StepsContainer Background { get; set; }
+
         [DataMember(Name = "rule")]
         public Rule Rule { get; set; }
-
-        
     }
 }
