+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/GherkinDocument.cs
@@ -1,19 +1,16 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-namespace Gherkin.Events.Args.Ast
+namespace Gherkin.CucumberMessages.Types
 {
-    public class GherkinDocumentEventArgs
+    public class GherkinDocument
     {
-        public GherkinDocumentEventArgs()
-        {
-            Comments = new List<Comment>();
-        }
+        [DataMember(Name = "uri")]
+        public string Uri { get; set; }
 
         [DataMember(Name = "feature")]
         public Feature Feature { get; set; }
-        [DataMember(Name = "uri")]
-        public string Uri { get; set; }
+
         [DataMember(Name = "comments")]
         public IReadOnlyCollection<Comment> Comments { get; set; }
     }
