+++ gherkin/dotnet/Gherkin/CucumberMessages/Types/Location.cs
@@ -1,0 +1,23 @@
+﻿using System.Runtime.Serialization;
+
+namespace Gherkin.CucumberMessages.Types
+{
+    public class Location
+    {
+        [DataMember(Name = "line")]
+        public int Line { get; set; }
+
+        [DataMember(Name = "column")]
+        public int Column { get; set; }
+
+        public Location()
+        {
+        }
+
+        public Location(int column, int line)
+        {
+            Column = column;
+            Line = line;
+        }
+    }
+}
