+++ gherkin/dotnet/Gherkin/Events/Args/Ast/Cell.cs
@@ -5,7 +5,7 @@ namespace Gherkin.Events.Args.Ast
     public class Cell
     {
         [DataMember(Name = "location")]
-        public Location Location { get; set; }
+        public CucumberMessages.Types.Location Location { get; set; }
         [DataMember(Name = "value")]
         public string Value { get; set; }
     }
