--- gherkin/dotnet/Gherkin/Events/Args/Ast/Location.cs
@@ -1,22 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace Gherkin.Events.Args.Ast
-{
-    public class Location
-    {
-        public Location(int column, int line)
-        {
-            Column = column;
-            Line = line;
-        }
-
-        public Location()
-        {
-        }
-
-        [DataMember(Name = "column")]
-        public int Column { get; set; }
-        [DataMember(Name = "line")]
-        public int Line { get; set; }
-    }
-}
